+++ magda-web-client/src/helpers/record.js
@@ -313,6 +313,7 @@ export function parseDistribution(
             numeric: numericFields
         };
     }
+
     return {
         identifier,
         title,
@@ -330,7 +331,8 @@ export function parseDistribution(
         visualizationInfo: aspects["visualization-info"]
             ? aspects["visualization-info"]
             : null,
-        compatiblePreviews
+        compatiblePreviews,
+        sourceDetails: aspects["source"]
     };
 }
 
@@ -425,7 +427,8 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
             isTimeSeries: visualizationInfo["timeseries"],
             chartFields,
             compatiblePreviews,
-            visualizationInfo: visualizationInfo ? visualizationInfo : null
+            visualizationInfo: visualizationInfo ? visualizationInfo : null,
+            sourceDetails: distributionAspects["source"]
         };
     });
     return {
@@ -443,6 +446,8 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
         publisher,
         error,
         linkedDataRating,
-        hasQuality
+        hasQuality,
+        sourceDetails: aspects["source"],
+        creation: datasetInfo["creation"]
     };
 }